Moved / rerouting to RedirectHome. Removed Layout. Cleaned protected routes
