--- src/layout/Layout.jsx
@@ -1,11 +1,0 @@
-import { Outlet } from 'react-router-dom';
-
-import './layout.css';
-
-export default function Layout() {
-	return (
-		<div className="layoutContainer">
-			<Outlet />
-		</div>
-	);
-}
