--- src/layout/layout.css
@@ -1,7 +1,0 @@
-.layoutContainer {
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	width: 100%;
-	height: 100dvh;
-}
